new method in multivariate instance tools that allows the splitting of a dataset with a proprtion.
For the new datasets created I split using 50%. train and test
and a seed of 0.

+++ TimeSeriesClassification/src/utilities/InstanceTools.java
@@ -154,7 +154,7 @@ public class InstanceTools {
  * @param propInTrain proportion of data for training
  * @return 
  */
-    public static Instances[] resampleInstances(Instances all, int seed, double propInTrain){
+    public static Instances[] resampleInstances(Instances all, long seed, double propInTrain){
         ClassDistribution classDist = new TreeSetClassDistribution(all);
         Map<Double, Instances> classBins = createClassInstancesMap(all);
